Plenty of changes to pass around the botuser id
Slack does not give us the bot user id beforehand, only when we connect
(start) to the RTM api. That said, this user id is used on messages
mentioning the bot, instead of something like @docbot, slack uses
`<@UTLA0123T>`.

Because of that, I had to change slightly the implementation to pass
around this value.

Initially I thought about setting a ENV var, but that would be worse for
testing, so I had to listen to my tests.